requirements/local.txt
-r base.txt

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.4
fabric==1.11.1